Create AURA-SIM900
This is code to manage a SIM900 mobile module http://www.seeedstudio.com/depot/gprs-shield-p-779.html. It receives telemetry produced by another Arduino via a WIRE I2C interface and sends it via HTTP puts to HABHUB. It also sends SMS messages, and is able to receive incoming calls.

The mobile module is manages via a state model. NOTE the ring indicator (RI) on the mobile module needs to be connected to pin A0 on the Arduino a as this provide confirmation of the state of the Mobile module. Please let me know if you find this useful..!